added exception listener tests
