+++ masonite/listeners/__init__.py
@@ -1,2 +1,1 @@
 from .BaseExceptionListener import BaseExceptionListener
-from .ExceptionListener import ExceptionListener
